soc: qcom: ipa: kill ipa_gsi_ep_config->ipa_if_aos
The ipa_if_aos field of the ipa_gsi_ep_config structure is assigned
for entries in the ipa_ep_configuration[] array.  These assignments
(I presume) match IPA/GSI configuration performed by the modem (and/or
Trust Zone), but seem to be only informational, because they are
never used in the Linux code.  Get rid of the field.

Signed-off-by: Alex Elder <elder@linaro.org>