soc: qcom: ipa: kill ipa_ep_configuration->valid
The valid field of the ipa_ep_configuration structure is always
assigned value true.  The only reason it might have been useful is
because the ipa_ep_configuration[] array is sparse, and there at one
time might have been code that iterated across all possible indexes
in that array.  But there is no such code now, so we can just get
rid of the valid field.

Signed-off-by: Alex Elder <elder@linaro.org>