soc: qcom: ipa: GSI transfer cleanup
The gsi_xfer_flag enumerated type has nothing to do with hardware.
It's possible that code previously used the particular values
assigned to the enumerated symbols to set register values, but that
is not the case now.  Change the numeric values assigned to these
symbols so they're more "normal."

Similarly, the gsi_xfer_elem structure is an abstracted form of
the hardware transfer element, but it isn't a hardware construct.
Change the type of the flags field to reflect the actual type it's
supposed to hold.

The values of the opcode in an ipahal_imm_cmd_pyld structure come
from an enumerated type, and if the type weren't hidden we would
represent it that way.  Still, add a comment to communicate this.

Signed-off-by: Alex Elder <elder@linaro.org>