soc: qcom: ipa: complete QMI INIT_DRIVER transaction
Since we are using a handler function for the modem's response to
the QMI INIT_DRIVER request, we need to set the result and complete
the transaction.  The transaction will be marked successful (result
is 0).

Note that qmi_txn_wait() won't time out, but it could return a
different error.  If the transaction actually completes, the return
value will be the result we set, 0, so any non-zero (negative)
result means we need to cancel the transaction.

Signed-off-by: Alex Elder <elder@linaro.org>