soc: qcom: ipa: kill "field_mask.h"
It turns out that there already exists macros for doing exactly what
I was doing in "field_mask.h".  The one missing feature was a macro
that produces the maximum representable value for a field, but I've
now defined that.  So replace the use of field_mask functions in the
IPA code with their equivalents from <linux/bitfield.h>:
    field_val(val, mask) -> FIELD_GET(mask, val)
    field_gen(val, mask) -> FIELD_PREP(mask, val)
    field_max(mask) -> FIELD_MAX(mask)
(Note that the arguments are reversed in the last two.)

And of course, remove "field_mask.h".

Signed-off-by: Alex Elder <elder@linaro.org>