soc: qcom: ipa: reverse Christmas tree
Srini mentioned something about David Miller preferring the "reverse
Christmas tree" style of automatic variable declaration.  I hadn't
heard of that but I can now say I've often used it, including in the
IPA code.  Anyway, small things like that affect one's sense of a
code's readability, and it's an easy thing to fix, so this patch
fixes it throughout the IPA code.  (It was a quick fix; I won't
guarantee I got them all.)

There are a few other odds and ends in here, but none should have
changed the behavior of the code.

Signed-off-by: Alex Elder <elder@linaro.org>