soc: qcom: ipa: kill ipahal_pkt_status->nat_type
The nat_type field of the ipahal_pkt_status structure is never used,
so get rid of it.  We still get back a nat_type value from the
hardware, but we'll just ignore it until we need it.  Get rid of
some other supporting code related to the nat_type field, including
the definition of the ipahal_pkt_status_nat_type enumerated type.

Signed-off-by: Alex Elder <elder@linaro.org>