soc: qcom: ipa: always record exception and status_opcode
In ipahal_pkt_status_parse(), if the status opcode or the exception
value is found to be invalid/undefined, it is not assigned to the
corresponding field in the ipahal_pkt_status structure.  This means
that field will have value 0, which is not well-defined.  In the
case of thee exception, it actually represents "no exception" which
is wrong.

Instead, *always* assign the values that the hardware supplies
regardless of whether they're valid, and let the caller take care of
validating them.

Signed-off-by: Alex Elder <elder@linaro.org>