soc: qcom: ipa: don't include opcode in payload
We don't need to include the opcode in the immediate command payload.
Replace calls to ipahal_imm_cmd_pyld_alloc() with calls to kzalloc()
and get rid of ipahal_imm_cmd_pyld_data().  Because we no longer
need to determine where the type-specific part of the payload is, we
can get rid of ipahal_imm_cmd_pyld_data() also.

Signed-off-by: Alex Elder <elder@linaro.org>