soc: qcom: ipa: assume hardware provides good masks
In several functions in "gsi.c" mask values are read to determine
which channel or event ring had an interrupt (or other event).
There is code that validates those values to ensure they don't
exceed the number of rings or channels the hardware original
reported was supported (either in channel_max or evt_ring_max).

The only reason the hardware would return a mask with a value that
was too big is if the hardware was broken.  But in general we must
assume the hardware is working correctly, so there's no reason we
should check for this particular type of bad behavior.

Convert these checks into assertions, to simplify the code.

Signed-off-by: Alex Elder <elder@linaro.org>