soc: qcom: ipa: bug fix in ipa_ap_suspend()
In ipa_ap_suspend(), each endpoint is checked to see if it is in
polling mode, and if so, suspend is prevented.  Note that currently
only consumer endpoints can be in polling mode, and ipa_ep_polling()
has an assertion checking that.  We need to be sure only to call it
from ipa_ep_polling() for consumer endpoints, to avoid failing the
assertion.

This was found during testing of suspend.

Signed-off-by: Alex Elder <elder@linaro.org>