soc: qcom: ipa: fix bug in ipahal_rt_generate_empty_img()
The size of the value "put" by put_unaligned() is determined by the
size of what the destination pointer refers to, *not* the size of
the value to be put.  Currently, ipahal_rt_generate_empty_img()
passes a void pointer, which is treated by the compiler to refer to
an object of size 1 byte, so we aren't actually writing what we
think we are when initializing tables.

The same bug exists in ipahal_flt_generate_empty_img()

Fix this bug in both places by passing a pointer to a 64-bit value
to put_unaligned() instead.

Signed-off-by: Alex Elder <elder@linaro.org>