soc: qcom: ipa: fix bug where tables are freed
It's an error to free the route and filter table after they have
been allocated.  They need to remain untouched until we're done
using them (and not be re-allocated for another purpose).  Stop
freeing these blocks of DMA memory after they're allocated.

NOTE that there is currently no cleanup function to undo this.  That
will come in some future updates, when we do a comprehensive scan
to fix that sort of thing.

Signed-off-by: Alex Elder <elder@linaro.org>