soc: qcom: ipa: tables are aligned
We know DMA allocations come back with at least 8 byte alignment,
and we know our table entries are 8 bytes each, so there's no need
for us to handle unaligned accesses when writing the route and
filter tables.  Switch these to use normal assignment for
initialization.  If table entry sizes ever got to be something odd
(which is unlikely) we might need to switch this back.

Signed-off-by: Alex Elder <elder@linaro.org>