arm64: dts: sdm845: kill qcom,rmnet-ipa
Rob Herring in review suggested that a node with nothing but a
compatible string usually indicates something that shouldn't be
there.  He's right.  We no longer support (or require) such a node
to exist in the DTB, so get rid of it in "sdm845.dtsi".  (The
binding document for this node was already eliminated.)

Signed-off-by: Alex Elder <elder@linaro.org>