soc: qcom: ipa: don't use ipa_ctx global in another spot
Use an argument passed to ipa_freeze_clock_vote_and_notify_modem()
rather than the global ipa_ctx.  This requires ipa_panic_notifier()
to have the pointer, but now that the notifier block is embedded in
the IPA structure it can derive that.  It also requires it to be
available in ipa_smp2p_modem_clk_query_isr(), but it's available
there via the dev_id argument.

Also add an IPA structure argument to ipa_reset_freeze_vote(), and
similarly use it rather than the global.  Currently that function
is unused but it will be needed once SSR support is added (soon).

Signed-off-by: Alex Elder <elder@linaro.org>