soc: qcom: ipa: remove initialized global
Remove the initialized global in "rmnet_ipa.c".  It is only set when
ipa_wwan_probe() has completed, which (assuming the "rmnet_ipa" node
is present in the DTB) will occur right after the driver is
registered.  And that occurs when ipa_wwan_init() is called.  The
flag is set to false when ipa_wwan_remove() is called.  This
currently doesn't happen, but would if SSR were supported.

Only ipa_post_init() calls ipa_wwan_init(), and it won't call it
twice.  So there's no point in having the initialized flag.

Signed-off-by: Alex Elder <elder@linaro.org>