soc: qcom: ipa: don't use global in rmnet_ipa suspend/resume
Pass the pointer returned by ipa_wwan_init() as an argument to the
rmnet_ipa suspend and resume functions, and use the argument rather
than the global variable to identify the rmnet_ipa_context structure
affected by the call.  Since ipa_wwan_resume() no longer needs to
adhere to the resume method interface, so give it a void return type.

Signed-off-by: Alex Elder <elder@linaro.org>