soc: qcom: ipa: eliminate rmnet_data device
The probe and exit (now begin and end) routines for the rmnet_data
interface never actually use their devices.  Nothing about it
requires an actual device, nor therefore a driver.

Replace the driver registration with a simple call to the begin
function, and the driver unregistration with a call to the end
function.  Pass NULL for the platform device pointers they are
provided (the only use of that was to print messages in the end
routine, and those prints are now gone).  Get rid of the driver and
the match table.

Signed-off-by: Alex Elder <elder@linaro.org>