soc: qcom: ipa: rename ipa_enable_clks()
Rename ipa_enable_clks() to be ipa_clock_enable(), and ipa_disable_clks()
to be ipa_clock_disable().  Have ipa_clock_enable() return a value,
which is honored when it's enabled explicitly the first time but
still ignored when called later on.

Get rid of the racy assertion in ipa_client_add_first(), and change the
logic there so the increment occurs regardless, but enabling the clock
and resuming pipes only occurs if the result of the increment is 1.

Signed-off-by: Alex Elder <elder@linaro.org>