soc: qcom: ipa: create post_init interlock
If the modem is doing initialization, ipa_post_init() will be called
sometime after the probe method returns.  The remove method needs to
know whether post_init has completed; we have a Boolean flag to
indicate this, but there's still a possibility that ipa_post_init()
could be executing at the time the remove method is called.

This patch creates an interlock to coordinate this.

First, if Trust Zone is doing the initialization, post_init will be
called and completed as part of the probe method.  The device lock
is held, and there is no chance for the remove method to be called
concurrently.  So there's no need to coordinate in this case.

If the modem is doing initialization ipa_post_init() will be called
as a result of an SMP2P interrupt.  In the remove method we start by
disabling the post_init IRQ.  We set a flag under protection of a
mutex to indicate we are shutting down.  That flag is checked in the
IRQ handler and the ipa_post_init() call is skipped if shutdown has
begun.  The mutex is taken before checking the flag, and is held for
the duration of the ipa_post_init() call.

This way we can ensure that the remove method will not proceed until
either ipa_post_init() is complete, or will not be run.

Signed-off-by: Alex Elder <elder@linaro.org>