soc: qcom: ipa: fix up GSI channel stats
Embed the stats structure inside a channel structure named "stats",
rather than having it be a parallel array.  Rename the type to be
"channel_stats".  Change the field names so they don't have "ch_"
prefixes.  Change the type of the fields to be u64.  Remove two
stats fields that are never referenced.

Signed-off-by: Alex Elder <elder@linaro.org>