soc: qcom: ipa: simplify global error handlers
All cases in gsi_isr_glob_chan_err() and gsi_isr_glob_evt_err()
print a message reporting the error code that was being reported.
These codes are hardware defined, so won't change, so we don't
really need to print a symbolic message in the event they occur.

In both functions, just use a common WARN() to print the code seen,
as well as the error EE value.  For OUT_OF_RESOURCES errors, we
complete anything waiting, and still log the condition with
dev_err().  If this occurs on an event ring, report the channel the
ring is associated with rather than the event ring id.

Signed-off-by: Alex Elder <elder@linaro.org>