soc: qcom: ipa: ignore command timeouts
If a GSI channel or event ring command times out, there's really not
much we can do.  It suggests the hardware is malfunctioning, and at
the moment we don't have any recourse (other than a kernel panic).

If a timeout occurs waiting for a command response, issue a WARN()
call so we can see what happened, but otherwise, ignore it.  This
removes a number of bug_on() calls, and gets rid of some other
conditional code as well.

One block of code, in gsi_channel_stop(), does more than the others
on a timeout.  If the command times out, it checks to see if the
state of the channel is stopped anyway, in which case the timeout is
ignored.  (In that case there's no harm in ignoring the timeout.)

Signed-off-by: Alex Elder <elder@linaro.org>