soc: qcom: ipa: get rid of some RX error handling
In ipa_wan_rx_pyld_hdlr() there is a check to endpoint source and
destination index and the packet size are all valid.  If they were
not valid, the hardware is broken, so this should effectively never
happen.  So get rid of the check.

In ipa_lan_rx_cb(), a similar check verifies the source index in the
status packet is in range, and that it represents an allocated
enpoint with a non-null notification callback.  The index must be in
range, or the hardware is broken.  And the other to tests should
never happen for correct code.  So get rid of this check too.

Signed-off-by: Alex Elder <elder@linaro.org>