soc: qcom: ipa: rename ipa_init_sram()
Rename ipa_init_sram() to be ipa_smem_init().  The use of "smem" is
consistent with the use of "smem_offset" and "smem_size" to describe
the shared portion of IPA memory, and the name now fits the
convention being used elsewhere.  Introduce ipa_smem_exit() as well,
and use it.  Currently it doesn't do anything.

Signed-off-by: Alex Elder <elder@linaro.org>