BACKPORT: UPSTREAM: net: dsa: Set the master device's MTU to account for DSA overheads
DSA tagging of frames sent over the master interface to the switch
increases the size of the frame. Such frames can then be bigger than
the normal MTU of the master interface, and it may drop them. Use the
overhead information from the tagger to set the MTU of the master
device to include this overhead.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit dc0fe7d47f9f2256e3aa695bb754625f37291e17
https://github.com/torvalds/linux.git)
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>