FROMLIST: clk: qcom: hfpll: ignore unused
When COMMON_CLK_DISABLED_UNUSED is set, in an effort to save power and
to keep the software model of the clock in line with reality, the
framework transverses the clock tree and disables those clocks that
were enabled by the firmware but have not been enabled by any device
driver.

If CPUFREQ is enabled, early during the system boot, cpufreq might
attempt to change the CPU frequency ("set_rate"). If the HFPLL is
selected as a provider, it will then change the rate for this clock.

When that happens and since the driver provides the "is_enabled"
callback, the "enable_count" flag will not be modified.

As boot continues, clk_disable_unused_subtree will run. And now,
since it wont find a valid counter in enable_count it will disable this
running clock therefore stopping the CPU.

The following commit fixes the above use case making sure the clock is
not disabled.

Co-developed-by: Niklas Cassel <niklas.cassel@linaro.org>
Signed-off-by: Niklas Cassel <niklas.cassel@linaro.org>
Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 42efe65efca7f767b41fc290feb484209f7d3cef)
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>