FROMGIT: regulator: fan53555: Add support for FAN53526
The FAN53526 differs from the FAN53555 only in that the mode bit in
VSEL0/VSEL1 is moved to the CONTROL register, the voltage selector mask
is extended by 1 bit and the step is different.

So extend the existing fan53555 driver to support FAN53526 as well.

BUG=b:124524751
TEST=Stepped through each frequency and observed voltage change

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>

FROM GIT:
https: //git.kernel.org/pub/scm/linux/kernel/git/broonie/regulator.git
commit-id: f2a9eb975ab2e848af68db26646793f5d9d097e8
Change-Id: Icf241c992c8c65812c2904a130fc45e3afbcf8f0