INTERNAL: arm64: dts: qcom: qcs404-mistral: Use buck for vdd_apc
In order to reach 1.4GHz the external FAN53526 is providing VDD_APC, os
override the default S3 from PMS405 as supply for the CPUs. The 1.4GHz
opp is defined and voltage values are adjusted to match valid voltage
steps of the FAN53526.

Finally the SPMI PMS405 S3 regulator is disabled.

BUG=b:124524751
TEST=Stepped through each frequency and observed voltage change

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Change-Id: Ifd1bb3cbe222c71d8b701f0ef861478ba0240d6c