saner calling conventions for put_mountpoint()/umount_mnt()
Instead of passing disposal list to put_mountpoint() (same for all but
one caller), rename it to __put_mountpoint() and add a single-argument
put_mountpoint() that would pass the default (&ex_mountpoints) to
__put_mountpoint().  Open-code the call of unhash_mount() in
mntput_no_expire() (where we do pass an unusual disposal list),
return the rest to single-argument variant.

To be folded into the earlier commits.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>